Auto-detect proportions of uploaded images
This makes a couple large changes to how the image data is passed
between Settings and ImageProcessor. imageWidth and imageHeight are
now data attributes of Settings that are bound via v-model
to the v-text-field elements, which are in turn passed down
as props to ImageProcessor. Image size detection is now done
in Settings right before the imageUrl is passed down to
ImageProcessor, but this necessarily entails loading the
image data from the objectURL twice, once to get the image size
to set it in Settings, and once to use the data in the draw call
in ImageProcessor. For loading large images this (might) have
some detrimental effect, so comments are welcome.

ImageProcessor now uses the props passed down to it from Settings
to determine the sizes used in the drawImage calls, as opposed
to looking at the value attribute of the v-text-field elements.

Some basic input validation is performed on imageHeight and
imageWidth in Settings with the new min{height, Width} data
attributes, which is bound to image{Width,Height}'s change
events, but I've found that the change events don't always act
fast enough to validate the values before they're sent down
to ImageProcessor and cause an exception. That being said, I
was thinking proper input validation was a task to be completed
in another branch and not this one.