Update README
Add a few instructions to flesh out the README a
bit, as well as a .gif that shows off some functionality.

It'd be good to make sure this has up-to-date screenshots/
videos of functionality, so people won't even have to
click on the app to see what it's about/how it works.