Change size inputs to sliders
Changes the v-text-fields that are used to input the mosaic height/width
to v-sliders, with text fields at the end indicating the exact value.

The height and width sliders should change in time with each other if
"preserve aspect ratio" is switched on.

Sliders necessarily enforce the min/max values, and the text box is able
to do bound-checking on the @change event, so this should properly
implement input validation, as well.