Fix image becoming squashed at small sizes
Since the aspect ratio was calculated every time the height/width
changed, the image would become square at small sizes, even when
Preserve Aspect Ratio was selected. This adds a new data property,
imageRatio, that is only set when the image is loaded from OpenImage and
when Preserve Aspect Ratio is unchecked. When Preserve Aspect Ratio is
checked, imageRatio will be used to determine how imageHeight and
imageWidth will respond to changes.